adjusted the growth timer and ordered things better
changed up the timer code to make it more scalable added an option to speed up growth by adding one of the mixing agents. Ordered around the scripts and prefabs so they are not in the same folder. Also made a little quest thingie that rewards with more seeds, it has no ui for now but ill work on that after unicorns lair.